+++ CHANGELOG.md
@@ -31,6 +31,7 @@ We used to have incremental versioning before `0.1.0`.
 - Forbids to use inconsistent `return` and `yield` statements
 - Forbids to use multiple `return` path in `try`/`expect`/`finally`
 - Forbids to use implicit string concatenation
+- Forbids to use useless `continue` nodes inside the loops
 
 ### Bugfixes